refactor: extract shared resource form code
